Added text above the player with a number that counts down showing how long the tag cooldown is for the player.
Added a new surface for drawing on front. The text above the player is blit on this surface so it will be drawn on the front. This surface is cleared after it has been blit on the display surface.

Fixed a bug where a player can tag multiple people if they collide with them at the same time. Now it selects the player with the most tag time (has been tagged the least) and if they are the same, it is random.

Made self.current_time in the player instead of using pygame.time for each method.