Made displaying the tag times also show the player keybinds. Removed show all text.
Added the keybinds for normal (4 players, commonly used), compact (8 players, fit more players, but OK) and max (16 players, for all the player sprites, each player gets 3 keys, which is a bit unconfortable but it is mean for multiple keyboards).
Temporarility rebinded keybinds for max players.
Aloiner added a map (unfinished).