Organised the settings into game, text and movement. The settings are now key value pairs of the name and a list of the value and a description. Updated the rest of the code to update the settings. The settings now work.
Removed show any text. Might add back.
Made proper labels above the players which move based on what is turned on. The labels move up if something under is turned on. Moved the 321 T lower. It also moves with the labels.

Aloiner's map updated.

All objects are players, and will be added in.