Changed the default body_text_offset of Button to 25 (instead of 30). Looks better, text closer.
Moved all the settings into one dictionary. That dictionary's key value pairs are: name, settings group. The settings group were the old settings.
Added the type in the settings group at index 2.

Made a settings page, where you can change the settings. On the top is the settings page selector that allow selection of the setting group. On the left are the options of the settings group, with individual settings. Settings can't be changed using this yet. The settings page is titled with Settings and the settings page.

Added a return page method in Game. It will go back to the menu if run, but not if it already on the menu or on the game.