Made list a called map names, which stores all the names of the maps. This list is used to open the dictionary assuming that the names of the files are: Map{name}.tmx
In settings, in each of the option details, more details were added.
True or false are left, int have a lower bound and upper bound, and text showing if that is a hard limit or not. For string, it is a list of options to cycle through.

Some moving of buttons.

Fixed pygame_gui.

Added a text box which is show if it a int being changed in settings, true and false buttons to change bools, and next and previous buttons to change strings.
The buttons are updated, text changed, then drawn.
The settings are changed when the buttons are pressed.