More parameters on draw text: - wrap_text which is to add a \n after all the full stops. - return_size which will return the text surface width and height.
Removed the button cooldown end and button cooldown from the buttons and only for the return button, in the main code. This is so only the quit button will have a cooldown so you don't quit after you press the return button in settings.

Made all the settings work. Now you can click a top page, the bottom will react with another list, and you can select the sub-options. Then to the right, it will display the title, the current state and the description. Depending on the type it will have a text box/buttons showing you the options. Warnings will display if you have entered something that might cause problems. The text will automatically update.

For floats, you can't type decimal points yet.