Reduced the bug where if there was a semi collidable under you and above you, then you would sink very slowly through the one above you, and it looked laggy.
Made semi_collision_full_rects, which are the full rects of semi_collision_sprites.
In Player update_touching_sides, added semi_top_rect 4 pixels under the top of the players, with a height of 1. If this rect is touching any of the semi_collision_full_rects, it will disable the semi collidable blocks.